renamed detailview in hierarchyview
