--- src/DetailView/HierarchyView.js
@@ -1,17 +1,0 @@
-import React, {Component} from "react";
-import {Card, CardText, CardTitle} from "material-ui";
-
-class HierarchyView extends Component {
-    render() {
-        return (
-            <Card>
-                <CardTitle title="EQMN model of hierarchy" />
-                <CardText>
-                    Model
-                </CardText>
-            </Card>
-        );
-    }
-}
-
-export default HierarchyView;
