+++ src/index.js
@@ -4,7 +4,7 @@ import {BrowserRouter as Router, Switch, Route} from "react-router-dom";
 import { css } from 'aphrodite';
 import App from "./App";
 import ListView from "./ListView/ListView";
-import DetailView from "./DetailView/DetailView";
+import HierarchyView from "./HierarchyView/HierarchyView";
 import NotFound from "./Utils/NotFound";
 import Header from "./Utils/Header";
 import {Container} from "react-grid-system";
@@ -21,8 +21,8 @@ ReactDOM.render(
                             exact path="/"
                             component={ListView}/>
                         <Route
-                            path="/details/:hierarchyId"
-                            component={DetailView}/>
+                            path="/hierarchy/:hierarchyId"
+                            component={HierarchyView}/>
                         <Route
                             path="*"
                             component={NotFound}/>
